+++ src/main/java/com/franjmr/restapikarate/ProceduresApplication.java
@@ -1,4 +1,4 @@
-package com.zara.procedures;
+package com.franjmr.restapikarate;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
