+++ src/main/java/com/franjmr/restapikarate/ServletInitializer.java
@@ -1,4 +1,4 @@
-package com.zara.procedures;
+package com.franjmr.restapikarate;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
