+++ src/main/java/com/franjmr/restapikarate/controllers/UpdateEmployeeWorkShiftController.java
@@ -1,8 +1,8 @@
-package com.zara.procedures.controllers;
+package com.franjmr.restapikarate.controllers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.zara.procedures.models.ProcedureUpdateWorkShift;
-import com.zara.procedures.utils.*;
+import com.franjmr.restapikarate.models.ProcedureUpdateWorkShift;
+import com.franjmr.restapikarate.utils.*;
 
 import javax.persistence.*;
 import javax.validation.Valid;
