+++ src/main/java/com/franjmr/restapikarate/models/ProcedureUpdateWorkShift.java
@@ -1,4 +1,4 @@
-package com.zara.procedures.models;
+package com.franjmr.restapikarate.models;
 
 import javax.persistence.*;
 
@@ -7,8 +7,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @Entity
 @NamedStoredProcedureQuery(
-    	name = "POL_ActualizacionTurno",
-        procedureName = "ENLACES_PKG_ZARA_META4.POL_ActualizacionTurno",
+    	name = "ActualizacionTurno",
+        procedureName = "PKG_FRANJRM.POL_ActualizacionTurno",
         parameters = {
           @StoredProcedureParameter(mode = ParameterMode.IN, name = "PCodigoEmpleado", type = String.class),
           @StoredProcedureParameter(mode = ParameterMode.IN, name = "PFechaInicio", type = String.class),
