+++ src/main/java/com/franjmr/restapikarate/services/EmployeeContactsService.java
@@ -1,4 +1,4 @@
-package com.zara.procedures.services;
+package com.franjmr.restapikarate.services;
 
 public interface EmployeeContactsService  {
 	String updateContacts();
