+++ src/test/java/com/zara/procedures/ProceduresApplicationTests.java
@@ -1,4 +1,4 @@
-package com.zara.procedures;
+package com.franjmr.restapikarate;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
